% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGInterScoreTest.R
\name{GGInterScoreTest}
\alias{GGInterScoreTest}
\title{Vector of test statistics for gene-gene interaction effects}
\usage{
GGInterScoreTest(X1, X2, Y, inter.mod = "continuous", perm.method = "parametric", 
                 N = 1000, U = NULL)
}
\arguments{
\item{X1}{a n x p1 matrix of genotypes (coded as 0/1/2)}

\item{X2}{a n x p2 matrix of genotypes (coded as 0/1/2)}

\item{Y}{a vector of length n containing the values of the binary phenotype (0/1)}

\item{inter.mod}{a character giving the chosen model for the gene-gene interaction: continuous (\code{inter.mod="continuous"}) or using dummy coding (\code{inter.mod="dummy"})}

\item{perm.method}{a character giving the chosen resampling method for the phenotype: paramatric bootstrap (\code{perm.method="parametric"}) or simple permutations (any other character)}

\item{N}{the number of permutations of the phenotype (default = 1000)}

\item{U}{an optional n x q matrix of covariates}
}
\value{
A list containing the following elements: 
\item{Z}{the vector of test statistics}
\item{Z0}{the matrix of permuted vectors of test statistics, displayed as rows}
\item{Sigma}{the correlation matrix corresponding to \code{Z}}
\item{Y0}{the matrix of permuted phenotypes, displayed as columns}
\item{inter.mod}{the chosen interaction model (\code{"continuous"} or \code{"dummy"})}
\item{X1}{the design matrix corresponding to the first gene}
\item{X2}{the design matrix corresponding to the second gene}
\item{S}{the design matrix for the interaction effects}
\item{W}{the design matrix for the main effects, containing an intercept, the covariates and the principal 
components for the main effects}
\item{undefined.inter}{the indices of the undefined interaction effects, (columns removed from \code{S})}
\item{ev}{the approximate eigendecomposition of \code{Sigma}}
}
\description{
\code{GGInterScoreTest} is used to compute the vector of score test statistics for testing the 
presence of an interaction effect between two genes on a binary phenotype using a continuous or
categorical modelling of SNPs.
}
